+++ internal/model/user.go
@@ -14,6 +14,6 @@ type UserClaims struct {
 }
 
 type Balance struct {
-	Current  int `json:"current"`
-	Withdraw int `json:"withdraw"`
+	Current  float64 `json:"current"`
+	Withdraw float64 `json:"withdraw"`
 }
